requirements.txt
certifi==0.0.8
chardet==1.0.1
dj-database-url==0.4.2
Django==1.11.6
django-cors-headers==2.1.0
djangorestframework==3.7.0
Flask==0.8
gunicorn==0.14.2
Jinja2==2.6
Pygments==2.2.0
pytz==2017.2
requests==0.11.1
Werkzeug==0.8.3
whitenoise==3.3.1
